Refactor: move global/organization permissions overview components into their own files
