+++ src/components/UI/Display/MAPI/permissions/PermissionsUseCases/UseCasesForCategory.tsx
@@ -1,10 +1,10 @@
 import { AccordionPanel, Box, Text } from 'grommet';
 import { IPermissionsUseCase } from 'MAPI/permissions/types';
-import React from 'react';
+import React, { useMemo } from 'react';
 import styled from 'styled-components';
 import UseCase, {
   Column,
-} from 'UI/Display/MAPI/permissions/PermissionsUseCase/UseCase';
+} from 'UI/Display/MAPI/permissions/PermissionsUseCases/UseCase';
 
 import UseCaseStatus from './UseCaseStatus';
 
@@ -15,7 +15,7 @@ const LabelText = styled(Text)`
 interface IUseCasesForCategoryProps {
   category: string;
   useCases: IPermissionsUseCase[];
-  statuses: Record<string, boolean | undefined>;
+  statuses: Record<string, Record<string, boolean>>;
   organizations?: IOrganization[];
   isSelected: boolean;
 }
@@ -27,6 +27,27 @@ const UseCasesForCategory: React.FC<IUseCasesForCategoryProps> = ({
   organizations,
   isSelected,
 }) => {
+  const categoryStatuses = useMemo(() => {
+    const statusesByNamespace: Record<string, boolean | undefined> = {};
+
+    const namespaces = organizations
+      ? organizations.map((org) => org.id)
+      : [''];
+
+    namespaces.forEach((namespace) => {
+      const values = useCases.map(
+        (useCase) => statuses?.[useCase.name][namespace]
+      );
+      statusesByNamespace[namespace] = values.every((v) => v === true)
+        ? true
+        : values.every((v) => v === false)
+        ? false
+        : undefined;
+    });
+
+    return statusesByNamespace;
+  }, [organizations, statuses, useCases]);
+
   return (
     <AccordionPanel
       header={
@@ -35,7 +56,10 @@ const UseCasesForCategory: React.FC<IUseCasesForCategoryProps> = ({
             <Box direction='row'>
               <LabelText margin={{ right: 'small' }}>{category}</LabelText>
               {!organizations && !isSelected && (
-                <UseCaseStatus value={statuses['']} displayText={false} />
+                <UseCaseStatus
+                  value={categoryStatuses['']}
+                  displayText={false}
+                />
               )}
             </Box>
           </Column>
@@ -43,7 +67,7 @@ const UseCasesForCategory: React.FC<IUseCasesForCategoryProps> = ({
             !isSelected &&
             organizations.map((org) => (
               <Column key={org.id}>
-                <UseCaseStatus value={statuses[org.id]} />
+                <UseCaseStatus value={categoryStatuses[org.id]} />
               </Column>
             ))}
         </Box>
@@ -53,7 +77,7 @@ const UseCasesForCategory: React.FC<IUseCasesForCategoryProps> = ({
         {useCases.map((useCase) => (
           <UseCase
             useCase={useCase}
-            statuses={statuses}
+            statuses={statuses[useCase.name]}
             organizations={organizations}
             key={useCase.name}
           />
